+++ features/support/env.rb
@@ -5,7 +5,7 @@ require 'rspec'
 require 'database_cleaner'
 require 'spinach/capybara'
 
-%w(gitolite_stub stubbed_repository valid_commit).each do |f|
+%w(namespaces_stub gitolite_stub stubbed_repository valid_commit).each do |f|
   require Rails.root.join('spec', 'support', f)
 end
