+++ spec/support/namespaces_stub.rb
@@ -1,0 +1,14 @@
+require 'namespace'
+require 'gitlab/project_mover'
+
+class Namespace
+  def ensure_dir_exist
+    true
+  end
+end
+
+class Gitlab::ProjectMover
+  def execute
+    true
+  end
+end
